+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 1.2.30
-- Disable "Block Useless Perks" in IQ Addons by default for new and existing users (if you turn it back on it will stay enabled)
+- Disable "Block Useless Perks" in IQ Addons by default for new and existing users (if you turn it back on it will stay enabled)<<<<<<< ours
 - Also disable Remove Reloading Screen's missing texture substitution on 26.2+ Vulkan to stop more crashes
+- Force disable wWaypoints' bannable "Stay Sneaked in Containers" option and block it in Exploit Preventer so it is safe to play with it on servers