+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 1.2.30
-- Disable "Block Useless Perks" in IQ Addons by default for new and existing users (if you turn it back on it will stay enabled)<<<<<<< ours
+- Disable "Block Useless Perks" in IQ Addons by default for new and existing users (if you turn it back on it will stay enabled)
 - Also disable Remove Reloading Screen's missing texture substitution on 26.2+ Vulkan to stop more crashes
 - Force disable wWaypoints' bannable "Stay Sneaked in Containers" option and block it in Exploit Preventer so it is safe to play with it on servers
+- PolyPlus now disables Minecraft's narrator while muted by default to avoid unexplained lagspikes (can be disabled in the config)