+++ CHANGELOG.md
@@ -1,2 +1,2 @@
-## Unreleased changes
+## 1.2.31
 - Fixed the RAM advisor sometimes telling you to lower your RAM allocation too much